data

+++ poker/src/main/scala/com/code/server/game/poker/paijiu/GamePaijiu100.scala
@@ -192,7 +192,7 @@ class GamePaijiu100 extends GamePaijiuCrazy {
     this.playerCardInfos.values.foreach(player=>{
       betCount += player.getBetNum()
     })
-    betCount>=this.roomPaijiu.bankerScore
+    betCount>=this.roomPaijiu.bankerScore.toInt
   }
